Big Rewrite #2
Rewrote the whole project. The script now loops through each cache files in the cache_files folder and updates the user_info.txt file with their followed artists, or if the user does not have their information saved it will do so. It then removes users whose cache file no longer appears in the cache folder to prevent unnecessary API calls. It then creates a list of all artists followed by all users so as to only call the API for each artist once. It collects the last 5 singles and albums for each artist, then creates a new music dict of albums/singles released in the past day. It then pairs the new music with users who follow the artist and updates all the necessary spotify playlists. Albums added to playlists per user are logged as a text file in the logs folder. Finally emails are sent out to users whose playlists were updated and who have an email on their account.